requirements.txt
pandas==2.2.1
pyarrow==15.0.2
numpy==1.26.4
fastapi==0.110.0
uvicorn==0.27.1
google-auth==2.27.0
google-auth-oauthlib==1.2.0
openpyxl
langchain==0.1.13
langchain-google-genai==0.0.9
langchain-google-vertexai==0.1.0
requests==2.31.0
chromadb==0.4.24
langchain-openai==0.0.8
langchain-community
google-cloud-secret-manager==2.19.0
python-multipart==0.0.9
google-api-python-client==2.122.0
google-cloud-firestore==2.16.0
ipython
google-cloud-storage==2.14.0
langfuse
rank_bm25